Split The WebIDL Binder to Its Own File
